requirements.txt
spotipy>=2.24.0
python-dotenv>=1.0.0
streamlit>=1.30.0
plotly>=5.18.0
